+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/handler/bridge/mock/MockBridgeDataConfigItem.java
@@ -45,7 +45,7 @@ public class MockBridgeDataConfigItem implements Bean {
 
     @Override
     public String toString() {
-        return "MockSourceDataConfigItem{" +
+        return "RealtimeMockSourceDataConfigItem{" +
                 "pointId=" + pointId +
                 ", pointType='" + pointType + '\'' +
                 '}';