+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/handler/source/mock/realtime/RealtimeMockSourceDataConfigItem.java
@@ -1,17 +1,17 @@
-package com.dwarfeng.fdr.impl.handler.source.mock;
+package com.dwarfeng.fdr.impl.handler.source.mock.realtime;
 
 import com.alibaba.fastjson.annotation.JSONField;
 import com.dwarfeng.subgrade.stack.bean.Bean;
 
 /**
- * 模拟数据配置项。
+ * 实时模拟数据配置项。
  *
  * @author DwArFeng
  * @since 2.0.0
  */
-public class MockSourceDataConfigItem implements Bean {
+public class RealtimeMockSourceDataConfigItem implements Bean {
 
-    private static final long serialVersionUID = 212897411936848123L;
+    private static final long serialVersionUID = -721550085312004643L;
 
     @JSONField(name = "point_id")
     private Long pointId;
@@ -19,10 +19,10 @@ public class MockSourceDataConfigItem implements Bean {
     @JSONField(name = "point_type")
     private String pointType;
 
-    public MockSourceDataConfigItem() {
+    public RealtimeMockSourceDataConfigItem() {
     }
 
-    public MockSourceDataConfigItem(Long pointId, String pointType) {
+    public RealtimeMockSourceDataConfigItem(Long pointId, String pointType) {
         this.pointId = pointId;
         this.pointType = pointType;
     }
@@ -45,7 +45,7 @@ public class MockSourceDataConfigItem implements Bean {
 
     @Override
     public String toString() {
-        return "MockSourceDataConfigItem{" +
+        return "RealtimeMockSourceDataConfigItem{" +
                 "pointId=" + pointId +
                 ", pointType='" + pointType + '\'' +
                 '}';