+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/handler/source/mock/realtime/RequiredPointType.java
@@ -1,4 +1,4 @@
-package com.dwarfeng.fdr.impl.handler.source.mock;
+package com.dwarfeng.fdr.impl.handler.source.mock.realtime;
 
 import java.lang.annotation.*;